#109 User restrictions with pundit

+++ srcs/app/controllers/api/users_controller.rb
@@ -6,8 +6,9 @@ module Api
   class UsersController < ApiController
     before_action :set_user,
                   only: %i[show update upload_avatar create_ignore destroy_ignore create_friendship destroy_friendship]
-    before_action :allowed?,
-                  only: %i[update upload_avatar create_ignore destroy_ignore create_friendship destroy_friendship]
+    # before_action :allowed?,
+    #              only: %i[update upload_avatar create_ignore destroy_ignore create_friendship destroy_friendship]
+    after_action :verify_authorized, except: %i[index show]
 
     UserReducer = Rack::Reducer.new(
       User.all.order(id: :asc),
@@ -22,7 +23,7 @@ module Api
     end
 
     def update
-      return render_not_allowed if user_params.key?(:banned) && current_user.admin? == false
+      authorize @user
 
       disconnect_banned_user(@user.id) if user_params.key?(:banned) && user_params.fetch(:banned) == true
 
@@ -35,34 +36,37 @@ module Api
     end
 
     def upload_avatar
+      authorize @user
       return render_error('Malformed', 422) unless params.key?(:avatar)
 
-      mini_image = MiniMagick::Image.new(params[:avatar].tempfile.path)
-      mini_image.resize '1200x1200'
-      @user.avatar.attach(params[:avatar])
+      attach_avatar
       url = url_for(@user.avatar)
       json_response({ image_url: url })
     end
 
     def create_ignore
+      authorize @user
       p = ignore_params
       UserIgnore.create!(user: @user, ignored_id: p[:ignored_id])
       json_response({ ignored_id: p[:ignored_id].to_i })
     end
 
     def destroy_ignore
+      authorize @user
       id = params.fetch(:ignored_id)
       UserIgnore.where(ignored_id: id, user: @user).destroy_all
       head :no_content
     end
 
     def create_friendship
+      authorize @user
       p = friendship_params
       Friendship.create!(friend_a: @user, friend_b_id: p[:friend_id])
       json_response({ friend_id: p[:friend_id].to_i })
     end
 
     def destroy_friendship
+      authorize @user
       id = params.fetch(:friend_id)
       Friendship.where('friend_a_id = ? or friend_b_id = ?', @user.id, @user.id).where(
         'friend_a_id = ? or friend_b_id = ?', id, id
@@ -71,12 +75,6 @@ module Api
     end
 
     private
-
-    def allowed?
-      return unless current_user.id != @user.id && current_user.admin? == false
-
-      render_not_allowed
-    end
 
     def ignore_params
       params.permit(:ignored_id)
@@ -87,11 +85,17 @@ module Api
     end
 
     def user_params
-      params.require(:user).permit(:two_factor, :nickname, :first_login, :banned)
+      params.require(:user).permit(policy(@user).permitted_attributes)
     end
 
     def set_user
       @user = User.find(params[:id])
+    end
+
+    def attach_avatar
+      mini_image = MiniMagick::Image.new(params[:avatar].tempfile.path)
+      mini_image.resize '1200x1200'
+      @user.avatar.attach(params[:avatar])
     end
   end
 end
